# guest: inlines host.GetConfig to config.Get