… examples/nodenumber/main.wasm
… guest/api/host.go