#### What type of PR is this?

/kind cleanup

#### What this PR does / why we need it:

This inlines `host.GetConfig()` to `config.Get()` after noticing logging will be in another package. This is simpler than describing a new term "host" which is only used to get configuration.


#### Which issue(s) this PR fixes:

NONE

#### Special notes for your reviewer:

This also does refactoring to sketch out how to test logic that requires wasm imports.

#### Does this PR introduce a user-facing change?

NONE

#### What are the benchmark results of this change?

```benchstat
goos: darwin
goarch: arm64
pkg: sigs.k8s.io/kube-scheduler-wasm-extension/internal/e2e/scheduler
                          │ before.txt  │              now.txt              │
                          │   sec/op    │   sec/op     vs base              │
Example_NodeNumber/New-12   73.68m ± 2%   75.34m ± 2%       ~ (p=0.093 n=6)
Example_NodeNumber/Run-12   46.12µ ± 4%   46.82µ ± 1%       ~ (p=0.394 n=6)
geomean                     1.843m        1.878m       +1.88%
```